ya elimina un post
