+++ src/firebase/autenticadorfirebase.js
@@ -1,11 +1,6 @@
 import { logInGoogle, logInFacebook, outUser } from './controladorfirebase.js'
 
- export const promAuthFace = () => {
-     logInFacebook()};
-       
- export const promAuthGoogle = () => {
-     logInGoogle()
- };
+ 
 
 
 export const promOutUser = () => {
@@ -16,14 +11,3 @@ export const promOutUser = () => {
   });
 }
 
- export const saveUsers = () => {
-    var user = firebase.auth().currentUser;
-
-    if (user != null) 
-    firebase.firestore().collection('usuarios').doc(user.uid).set({
-      name : user.displayName,
-      email : user.email,
-      photoUrl : user.photoURL,
-      uid : user.uid
-    })
- };
